LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/background-position-composition-expected.txt


FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (-0.25) should be [75px 275px, 75px 275px] assert_equals: expected "75px 275px , 75px 275px " but got "75px 275px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (0) should be [100px 300px, 100px 300px] assert_equals: expected "100px 300px , 100px 300px " but got "100px 300px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (0.25) should be [125px 325px, 125px 325px] assert_equals: expected "125px 325px , 125px 325px " but got "125px 325px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (0.5) should be [150px 350px, 150px 350px] assert_equals: expected "150px 350px , 150px 350px " but got "150px 350px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (0.75) should be [175px 375px, 175px 375px] assert_equals: expected "175px 375px , 175px 375px " but got "175px 375px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (1) should be [200px 400px, 200px 400px] assert_equals: expected "200px 400px , 200px 400px " but got "200px 400px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to add [160px 260px] at (1.25) should be [225px 425px, 225px 425px] assert_equals: expected "225px 425px , 225px 425px " but got "225px 425px "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (-0.25) should be [110% 105%, 110% 105%] assert_equals: expected "110 % 105 % , 110 % 105 % " but got "110 % 105 % "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0) should be [100% 100%, 100% 100%] assert_equals: expected "100 % 100 % , 100 % 100 % " but got "100 % 100 % "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0.25) should be [90% 95%, 90% 95%] assert_equals: expected "90 % 95 % , 90 % 95 % " but got "90 % 95 % "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0.5) should be [80% 90%, 80% 90%] assert_equals: expected "80 % 90 % , 80 % 90 % " but got "80 % 90 % "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (0.75) should be [70% 85%, 70% 85%] assert_equals: expected "70 % 85 % , 70 % 85 % " but got "70 % 85 % "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (1) should be [60% 80%, 60% 80%] assert_equals: expected "60 % 80 % , 60 % 80 % " but got "60 % 80 % "
FAIL Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%] to add [right 80% bottom 40%] at (1.25) should be [50% 75%, 50% 75%] assert_equals: expected "50 % 75 % , 50 % 75 % " but got "50 % 75 % "
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (-0.25) should be [110% 105%, 235% 230%]
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (0) should be [100% 100%, 200% 200%]
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (0.25) should be [90% 95%, 165% 170%]
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (0.5) should be [80% 90%, 130% 140%]
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (0.75) should be [70% 85%, 95% 110%]
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (1) should be [60% 80%, 60% 80%]
PASS Compositing: property <background-position> underlying [top 20% left 40%] from add [left 60% top 80%, top 180% left 160%] to add [right 80% bottom 40%] at (1.25) should be [50% 75%, 25% 50%]
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (-0.25) should be [75px 150px, 75px 150px] assert_equals: expected "75px 150px , 75px 150px " but got "75px 150px "
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (0) should be [100px 200px, 100px 200px] assert_equals: expected "100px 200px , 100px 200px " but got "100px 200px "
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (0.25) should be [125px 250px, 125px 250px] assert_equals: expected "125px 250px , 125px 250px " but got "125px 250px "
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (0.5) should be [150px 300px, 150px 300px] assert_equals: expected "150px 300px , 150px 300px " but got "150px 300px "
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (0.75) should be [175px 350px, 175px 350px] assert_equals: expected "175px 350px , 175px 350px " but got "175px 350px "
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (1) should be [200px 400px, 200px 400px] assert_equals: expected "200px 400px , 200px 400px " but got "200px 400px "
FAIL Compositing: property <background-position> underlying [40px 140px] from replace [100px 200px] to add [160px 260px] at (1.25) should be [225px 450px, 225px 450px] assert_equals: expected "225px 450px , 225px 450px " but got "225px 450px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (-0.25) should be [75px 275px, 75px 275px] assert_equals: expected "75px 275px , 75px 275px " but got "75px 275px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (0) should be [100px 300px, 100px 300px] assert_equals: expected "100px 300px , 100px 300px " but got "100px 300px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (0.25) should be [125px 325px, 125px 325px] assert_equals: expected "125px 325px , 125px 325px " but got "125px 325px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (0.5) should be [150px 350px, 150px 350px] assert_equals: expected "150px 350px , 150px 350px " but got "150px 350px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (0.75) should be [175px 375px, 175px 375px] assert_equals: expected "175px 375px , 175px 375px " but got "175px 375px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (1) should be [200px 400px, 200px 400px] assert_equals: expected "200px 400px , 200px 400px " but got "200px 400px "
FAIL Compositing: property <background-position> underlying [40px 140px] from add [60px 160px] to replace [200px 400px] at (1.25) should be [225px 425px, 225px 425px] assert_equals: expected "225px 425px , 225px 425px " but got "225px 425px "
